chore: Add SLURM partition_queue.sh script for querying job queue
A new script, `partition_queue.sh`, has been added to the SLURM directory. This script allows users to query the job queue for the `epyc-gpu-test` partition. It provides a convenient way to check the status of jobs in this specific partition. This addition enhances the functionality of the SLURM scripts and improves the user experience.