chore: Update SLURM partition_queue.sh script to allow overriding the partition name
The SLURM partition_queue.sh script has been updated to allow overriding the default partition name via a command line argument. This enhancement provides flexibility for users to query job queues in different partitions without modifying the script. It improves the usability of the script and enhances the user experience.