chore: Add script to set email as environment variable
A new script, `set_email_env_variable.sh`, has been added to the SLURM directory. This script sets the email as an environment variable in the `.bashrc` file if it is not already set. It provides a convenient way for users to ensure that the SLURM_EMAIL variable is properly configured. This addition improves the usability of the SLURM scripts and enhances the user experience.